migration: Converted internal tests to break import loop
Importing the dummy provider in the internal test module caused an
import loop. Converting the internal tests to external tests fixes the
issue.
