+++ migration/export_test.go
@@ -1,0 +1,10 @@
+// Copyright 2016 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package migration
+
+var (
+	ControllerValues         = controllerValues
+	UpdateConfigFromProvider = updateConfigFromProvider
+	GetCharmStoragePath      = getCharmStoragePath
+)
